+++ .eslintrc.js
@@ -9,6 +9,7 @@ module.exports = {
     node: true,
   },
   rules: {
+    'jest/no-test-return-statement': 0,
     'implicit-arrow-linebreak': 0,
     'no-param-reassign': 0,
     'no-underscore-dangle': 0,